move count_parameters
